tests/wpt/web-platform-tests/offscreen-canvas/the-offscreen-canvas/context.casesensitive.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: context.casesensitive</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>context.casesensitive</h1>
<p class="desc">Context name "2D" is unrecognised; matching is case sensitive</p>


<script>
var t = async_test("Context name \"2D\" is unrecognised; matching is case sensitive");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

var offscreenCanvas2 = new OffscreenCanvas(100, 50);
assert_throws(new TypeError(), function() { offscreenCanvas2.getContext('2D'); });

t.done();

});
</script>
